smartdatapay-predev/langchain/requirements.txt
python-dotenv
# gpt_index==0.4.24
langchain==0.0.148
# langchain
# llama-index==0.5.6
llama-index
gradio
ipython
